updating nix files (#99)

+++ requirements.nix
@@ -1,8 +1,8 @@
-# generated using pypi2nix tool (version: 1.5.0)
+# generated using pypi2nix tool (version: 1.8.0)
 # See more at: https://github.com/garbas/pypi2nix
 #
 # COMMAND:
-#   pypi2nix -V 3.5 -r requirements.txt
+#   pypi2nix -v -V 3.5 -r requirements.txt
 #
 
 { pkgs ? import <nixpkgs> {}
@@ -13,11 +13,11 @@ let
   inherit (pkgs) makeWrapper;
   inherit (pkgs.stdenv.lib) fix' extends inNixShell;
 
-  pythonPackages = import "${toString pkgs.path}/pkgs/top-level/python-packages.nix" {
+  pythonPackages =
+  import "${toString pkgs.path}/pkgs/top-level/python-packages.nix" {
     inherit pkgs;
     inherit (pkgs) stdenv;
     python = pkgs.python35;
-    self = pythonPackages;
   };
 
   commonBuildInputs = [];
@@ -31,11 +31,13 @@ let
         buildInputs = [ makeWrapper ] ++ (builtins.attrValues pkgs);
         buildCommand = ''
           mkdir -p $out/bin
-          ln -s ${pythonPackages.python.interpreter} $out/bin/${pythonPackages.python.executable}
-          for dep in ${builtins.concatStringsSep " " (builtins.attrValues pkgs)}; do
+          ln -s ${pythonPackages.python.interpreter}               $out/bin/${pythonPackages.python.executable}
+          for dep in ${builtins.concatStringsSep " "               (builtins.attrValues pkgs)}; do
             if [ -d "$dep/bin" ]; then
               for prog in "$dep/bin/"*; do
-                ln -s $prog $out/bin/`basename $prog`
+                if [ -f $prog ]; then
+                  ln -s $prog $out/bin/`basename $prog`
+                fi
               done
             fi
           done
@@ -61,7 +63,97 @@ let
 
   python = withPackages {};
 
-  generated = import ./requirements_generated.nix { inherit pkgs python commonBuildInputs commonDoCheck; };
-  overrides = import ./requirements_override.nix { inherit pkgs python; };
+  generated = self: {
 
-in python.withPackages (fix' (extends overrides generated))
+    "GitPython" = python.mkDerivation {
+      name = "GitPython-2.1.5";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/6d/c1/d1c852f787a1ad4fd8346603e520e2b8b886158f177276cd8c2acf594001/GitPython-2.1.5.tar.gz"; sha256 = "5c00cbd256e2b1d039381d4f7d71fcb7ee5cc196ca10c101ff7191bd82ab5d9c"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [
+      self."gitdb2"
+    ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.bsdOriginal;
+        description = "Python Git Library";
+      };
+    };
+
+
+
+    "Jinja2" = python.mkDerivation {
+      name = "Jinja2-2.9.6";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/90/61/f820ff0076a2599dd39406dcb858ecb239438c02ce706c8e91131ab9c7f1/Jinja2-2.9.6.tar.gz"; sha256 = "ddaa01a212cd6d641401cb01b605f4a4d9f37bfc93043d7f760ec70fb99ff9ff"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [
+      self."MarkupSafe"
+    ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.bsdOriginal;
+        description = "A small but fast and easy to use stand-alone template engine written in pure python.";
+      };
+    };
+
+
+
+    "MarkupSafe" = python.mkDerivation {
+      name = "MarkupSafe-1.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/4d/de/32d741db316d8fdb7680822dd37001ef7a448255de9699ab4bfcbdf4172b/MarkupSafe-1.0.tar.gz"; sha256 = "a6be69091dac236ea9c6bc7d012beab42010fa914c459791d627dad4910eb665"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.bsdOriginal;
+        description = "Implements a XML/HTML/XHTML Markup safe string for Python";
+      };
+    };
+
+
+
+    "gitdb2" = python.mkDerivation {
+      name = "gitdb2-2.0.2";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/be/eb/69f956a2b4b7c529999e624ce86d7a986a29b23f15599e8e58e17ffd9d44/gitdb2-2.0.2.tar.gz"; sha256 = "f2e36d7561e91f30a6a44858756dc020d8f1e81ca6e4185979d5c6c24c648070"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [
+      self."smmap2"
+    ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.bsdOriginal;
+        description = "Git Object Database";
+      };
+    };
+
+
+
+    "smmap2" = python.mkDerivation {
+      name = "smmap2-2.0.3";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/48/d8/25d9b4b875ab3c2400ec7794ceda8093b51101a9d784da608bf65ab5f5f5/smmap2-2.0.3.tar.gz"; sha256 = "c7530db63f15f09f8251094b22091298e82bf6c699a6b8344aaaef3f2e1276c3"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [ ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "";
+        license = licenses.bsdOriginal;
+        description = "A pure python implementation of a sliding window memory map manager";
+      };
+    };
+
+  };
+  overrides = import ./requirements_override.nix { inherit pkgs python; };
+  commonOverrides = [
+
+  ];
+
+in python.withPackages
+   (fix' (pkgs.lib.fold
+            extends
+            generated
+            ([overrides] ++ commonOverrides)
+         )
+   )
